refactor: 이미 반영된 Cart -> cart 폴더의 싱크 이슈로 인한 재정비
