+++ src/refactoring/components/CartPage.tsx
@@ -1,6 +1,6 @@
-import { CartDetails, ProductList } from "@/refactoring/components/cart";
-
 import { Layout } from "@/refactoring/components/shared";
+import ProductList from "./cart/ProductList";
+import CartDetails from "./cart/details/CartDetails";
 
 export const CartPage = () => {
   return (
@@ -8,7 +8,6 @@ export const CartPage = () => {
       <Layout.Title>장바구니</Layout.Title>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
         <ProductList />
-
         <CartDetails />
       </div>
     </Layout.Container>
